@@ -5,7 +5,7 @@
 </template>
 
 <script>
-const axios = require("axios").default;
+import axios from "axios";
 
 export default {
   name: "ServerLogView",
@@ -24,7 +24,7 @@ export default {
     try {
       const response = await axios.get("/api/logs/get", { params: {} });
       this.log_lines = response.data;
-      document.title = "WhatRecord? Server logs";
+      document.title = "whatrecord? Server logs";
     } catch (error) {
       console.error(error);
     }
