@@ -16,7 +16,7 @@
 </template>
 
 <script>
-import DictionaryTable from "./dictionary-table.vue";
+import DictionaryTable from "@/components/dictionary-table.vue";
 
 export default {
   name: "AsynPort",
@@ -27,8 +27,7 @@ export default {
   beforeCreate() {
     // TODO: I don't think this is circular; why am I running into this?
     // V2 ref: https://vuejs.org/v2/guide/components-edge-cases.html#Circular-References-Between-Components
-    this.$options.components.DictionaryTable =
-      require("./dictionary-table.vue").default;
+    this.$options.components.DictionaryTable = DictionaryTable;
   },
 };
 </script>
