@@ -207,7 +207,7 @@ export default {
   computed: {
     appliance_viewer_url() {
       const appliance_viewer_url =
-        process.env.VUE_APP_WHATRECORD_ARCHIVER_URL || "";
+        import.meta.env.WHATRECORD_ARCHIVER_URL || "";
       if (!appliance_viewer_url || !this.record) {
         return null;
       }
