@@ -51,7 +51,7 @@
       :globalFilterFields="global_filter_fields"
     >
       <template #header>
-        <div class="p-d-flex p-jc-between">
+        <div class="flex justify-content-between">
           <MultiSelect
             :modelValue="selected_columns"
             :options="columns"
@@ -270,7 +270,7 @@ export default {
     }),
   },
   created() {
-    document.title = `WhatRecord? happi plugin`;
+    document.title = `whatrecord? happi plugin`;
     this.init_filters();
   },
   mounted() {
