@@ -20,7 +20,7 @@
       :globalFilterFields="global_filter_fields"
     >
       <template #header>
-        <div class="p-d-flex p-jc-between">
+        <div class="flex justify-content-between">
           <MultiSelect
             :modelValue="selected_columns"
             :options="columns"
@@ -162,7 +162,7 @@ export default {
     }),
   },
   created() {
-    document.title = `WhatRecord? netconfig ${this.item_name}`;
+    document.title = `whatrecord? netconfig ${this.item_name}`;
     this.init_filters();
   },
   mounted() {
